reverting change to return strings
